% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_message.R
\name{create_message}
\alias{create_message}
\title{Write the message}
\usage{
create_message(image, note = "Boo")
}
\arguments{
\item{image}{A ggplot object.}

\item{note}{A text string.}
}
\description{
Write your orange message in the sky.
}
