% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_jackolantern.R
\name{draw_jackolantern}
\alias{draw_jackolantern}
\title{Carve your jack-o-lantern}
\usage{
draw_jackolantern(message = "Happy Autumn!", shape = "face")
}
\arguments{
\item{message}{A text string.}

\item{shape}{The image to carve on your pumpkin. Currently accepts "face"
and "cat".}
}
\description{
Carve your jack-o-lantern
}
\examples{

pumpkin <- draw_jackolantern()
print(pumpkin)



}
